kernel: tick-sched: Fix compilation error on 32-bit
do_div() on 32-bit is changed into __div64_32(uint64_t *n, uint32_t
base), which requires its operand type to be specified.
Update the variable type to fix the compilation issue.

+++ kernel/time/tick-sched.c
@@ -1165,7 +1165,7 @@ void tick_irq_enter(void)
 static void update_rq_stats(void)
 {
 	unsigned long jiffy_gap = 0;
-	unsigned int rq_avg = 0;
+	unsigned long long rq_avg = 0;
 	unsigned long flags = 0;
 
 	jiffy_gap = jiffies - rq_info.rq_poll_last_jiffy;
